lint fixes, check exists before clearing

+++ tasks/initialize.js
@@ -3,20 +3,22 @@ const async = require('async');
 const ClientKitTask = require('clientkit-task');
 const mkdirp = require('mkdirp');
 const rmdir = require('rmdir');
+const fs = require('fs');
 
 class InitTask extends ClientKitTask {
   process(distDir, whatever, processDone) {
     async.auto({
       clearDist: (done) => {
-        rmdir(distDir, done);
+        if (fs.existsSync(distDir)) {
+          return rmdir(distDir, done);
+        }
+        done();
       },
       mkdist: ['clearDist', (results, done) => {
         mkdirp(distDir);
         done();
       }],
-    }, () => {
-      return processDone();
-    });
+    }, processDone);
   }
 }
 module.exports = InitTask;
